move useUserTenantPermissions hook back to hooks folder
