--- src/tmp/hooks/index.js
@@ -1,1 +1,0 @@
-export { default as useUserTenantPermissions } from './useUserTenantPermissions';
